requirements.txt
numpy
pandas
tensorrt
torch
pycuda
pyctcdecode
transformers
datasets
evaluate
hf_transfer
kenlm @ git+https://github.com/kpu/kenlm.git@master
jiwer
accelerate
# transformer_engine @ git+https://github.com/NVIDIA/TransformerEngine.git@stable
scikit-learn
pytorch_optimizer
librosa
soundfile
safetensors
tensorboard
psutil
opencv-python
pycocotools
onnxruntime-gpu
ensemble_boxes